% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_vars.R
\name{drop_vars}
\alias{drop_vars}
\title{Drop one or more variables from a data frame}
\usage{
drop_vars(input_data, ..., .strict = FALSE, verbose = FALSE)
}
\arguments{
\item{\dots}{passed to [tidyselect::vars_select()]}

\item{.strict}{logical}

\item{verbose}{logical}
}
\description{
Drop one or more variables from a data frame
}
\examples{
df <- data_frame(foo = 1, bar = 2)
drop_vars(df, foo)
drop_vars(df, apple)

}
