% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_var.R
\name{find_var}
\alias{find_var}
\title{find_var}
\usage{
find_var(input_data, suffix)
}
\arguments{
\item{input_data}{tabular data, like a tibble}

\item{suffix}{character}
}
\value{
First such variable (issues warning if more than one found)
}
\description{
Find a single variable, with a given suffix, among the variables in your data
}
