% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_vars.R, R/find_id_var.R, R/find_qty_var.R,
%   R/find_year_var.R
\name{find_vars}
\alias{find_vars}
\alias{find_all_vars}
\alias{find_var}
\alias{find_id_var}
\alias{find_qty_var}
\alias{find_year_var}
\title{Find matching column names}
\usage{
find_all_vars(...)

find_var(...)

find_id_var(input_data, ...)

find_qty_var(input_data, ...)

find_year_var(input_data, ...)
}
\arguments{
\item{input_data}{(tabular data) must have one or more columns}

\item{verbose}{(logical) display messages}

\item{pattern}{(character) regular expression (also see [glob2rx()])}
}
\value{
(character) name of variable(s) matching `pattern`.
}
\description{
Find the names of variable(s) matching a given pattern. If there are no matches,
display an error message and stop.
}
\details{
`find_var()` stops if more than one variable in `input_data` matches `pattern`.
}
\section{Functions}{
\itemize{
\item \code{find_all_vars}: Find one or more

\item \code{find_var}: Find exactly one

\item \code{find_id_var}: Find variable ending in "_id"

\item \code{find_qty_var}: Find variable ending in "qty"

\item \code{find_year_var}: Find variable named "year" or ending in "yr"
}}

