% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_vars.R
\name{drop_vars}
\alias{drop_vars}
\title{Drop one or more variables from a data frame}
\usage{
drop_vars(input_data, ..., .strict = FALSE, verbose = TRUE)
}
\arguments{
\item{input_data}{tabular data}

\item{\dots}{passed to [tidyselect::vars_select()]}

\item{.strict}{logical}

\item{verbose}{logical}
}
\description{
Drop one or more variables from a data frame
}
\examples{
drop_vars(mtcars, "cyl")
drop_vars(mtcars, c("mpg", "cyl"))
drop_vars(mtcars, mpg, cyl)
drop_vars(mtcars, mpg, foo, bar)
drop_vars(mtcars, c("mpg", "foo", "bar"))
v1 <- c("mpg", "cyl"); drop_vars(mtcars, !!v1)
v1 <- c("mpg", "cyl"); drop_vars(mtcars, !!!v1)
v2 <- c("mpg", "foo", "bar"); drop_vars(mtcars, !!!v2)

}
