% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_var.R, R/find_qty_var.R, R/find_year_var.R
\name{find_var}
\alias{find_var}
\alias{find_qty_var}
\alias{find_year_var}
\title{find_var}
\usage{
find_var(input_data, pattern, verbose = getOption("verbose"))

find_qty_var(...)

find_year_var(...)
}
\arguments{
\item{input_data}{tabular data, like a tibble}

\item{pattern}{regular expression (also see [glob2rx()])}

\item{verbose}{display messages}
}
\value{
(character) name of variable matching `pattern`.
}
\description{
Find the variable (i.e., column) matching a given pattern.
}
\details{
Warns if more than one variable in `input_data` matches `pattern`,
  but always returns the first such variable found. If there are no matches,
  `find_var()` will display an error message and stop.
}
\section{Functions}{
\itemize{
\item \code{find_qty_var}: Find variable ending in "qty"

\item \code{find_year_var}: Find variable named "year" or ending in "yr"
}}

